% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{bin.genes.genomic.regions}
\alias{bin.genes.genomic.regions}
\title{Binning genes into genomic windows.
\code{bin.genes.genomic.regions} Adds an extra colum \code{bin.tag} grouping genes to a genomic bins defined by an input \code{window}}
\usage{
bin.genes.genomic.regions(gene.pos, window = 5e+05)
}
\arguments{
\item{gene.pos}{Data.frame with "geneid" (ensembl ids), "chr" (only somatic, and as integers (1:22)), "left"(start), "right" (end). For reference please look at data(gene.info).}

\item{window}{Number of bases to define a genomic loci, 1MB (1000000) is set as defualt}
}
\value{
A dataframe with \code{gene.set}, its genomic position and an extra colum \code{loci} grouping genes to a genomic loci
}
\description{
Binning genes into genomic windows.
\code{bin.genes.genomic.regions} Adds an extra colum \code{bin.tag} grouping genes to a genomic bins defined by an input \code{window}
}
\examples{
gene.pos.test.bin <- bin.genes.genomic.regions(gene.pos.test, window=1000000)

}
