% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{col.data}
\alias{col.data}
\title{Sample information from CD4+ T cells from healthy and Celiac patients from GSE69549}
\format{A data.frame with 74 rows and 4 variables}
\usage{
data(col.data)
}
\description{
A data.frame containing sample information from GSE69549
}
\keyword{datasets,}
\keyword{information}
\keyword{sample}
