% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{convert_types}
\alias{convert_types}
\title{Converts data types}
\usage{
convert_types(colClasses)
}
\arguments{
\item{colClasses}{A character vector specifying the classes of the columns. This vector is used by (\code{\link{read.table}}).}
}
\value{
A character vector vector specifying the classes of the columns. This vector is used by (\code{\link[readxl]{read_excel}}).
}
\description{
Converts the character names of the R standards types to types which can be used by (\code{\link[readxl]{read_excel}}).
}
