% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{count.table}
\alias{count.table}
\title{Gene expression from CD4+ T cells from healthy and Celiac patients from GSE69549}
\format{A data.frame with 23228 genes (by rows) and 74 samples (by column)}
\usage{
data(count.table)
}
\description{
A data.frame containing RNASeq based expression levels from GSE69549
}
\keyword{datasets,}
\keyword{expression}
