% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{gene.info}
\alias{gene.info}
\title{Gene information}
\format{A data.frame with 74 rows and 4 variables}
\usage{
data(gene.info)
}
\description{
A data.frame with genomic positions for each gene version GRCh37.75
}
\keyword{datasets,}
\keyword{information}
\keyword{sample}
