% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{get.genes.within.loci}
\alias{get.genes.within.loci}
\title{get.genes.within.loci
\code{get.genes.within.loci} Given a set of \code{positions}(genomic coordinates), will return a data.frame with the genes within the input \code{positions}}
\usage{
get.genes.within.loci(positions, position.colnames = c("CHR", "BP",
  "SNP"), gene.window = 5e+05, gene.info, pVal.col = NULL)
}
\arguments{
\item{positions}{A data.frame with genomic coordinates.}

\item{position.colnames}{Character vector defining the colnames used for defining the genomic coordinates within the \code{positions} data.frame.}

\item{gene.window}{Numeric, which will deifine the genomic window that will be used to get neighbouring genes given a certain genomic position defined by \code{positions}.}

\item{gene.info}{A data.frame with genomic coordinates and genes. For reference please look at data(gene.info)}

\item{pVal.col}{Character, if provided, then an extra column to the returned data-frame will be added with the pValue of the SNP associated.}
}
\value{
A dataframe
}
\description{
get.genes.within.loci
\code{get.genes.within.loci} Given a set of \code{positions}(genomic coordinates), will return a data.frame with the genes within the input \code{positions}
}
