% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{get.sampled.genes}
\alias{get.sampled.genes}
\title{get.sampled.genes
\code{get.sampled.genes} Given a set of \code{input.genes} (which should be in enemsebl) this function wil return a list of \code{n.genes.sets} of genes which represent the same number of genomic regions as the \code{input.genes}}
\usage{
get.sampled.genes(input.genes, gene.info.binned, binning.widow = 125000,
  n.excess.genes.limit = 10, n.genes.sets = 1000, parallel = TRUE,
  n.clusters = 3)
}
\arguments{
\item{input.genes}{Character vector of ensembl ids}

\item{binning.widow}{Numeric, window size which will define the genomic regions and bining the genes. Default is 125 Kb to each side.}

\item{n.genes.sets}{Number of random gene sets that are going to be generated.}

\item{parallel}{Run the sampling iterations in parallel. At the moment this is the only available option.}

\item{gene.info}{A data.frame object with the genomic coordinates of the genes used in the study.}
}
\value{
A dataframe with \code{gene.set}, its genomic position and an extra colum \code{loci} grouping genes to a genomic loci
}
\description{
get.sampled.genes
\code{get.sampled.genes} Given a set of \code{input.genes} (which should be in enemsebl) this function wil return a list of \code{n.genes.sets} of genes which represent the same number of genomic regions as the \code{input.genes}
}
