% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{GREA.test}
\alias{GREA.test}
\title{GREA.test
\code{GREA.test} plost the ditribution and observed overlap}
\usage{
GREA.test(sampled.genes.sets, query.gene.set, genomic.region.gene.sets,
  plot = TRUE)
}
\arguments{
\item{sampled.genes.sets}{A list where each element is a character vector of ensembl IDs, from the \code{get.sampled.genes} function}

\item{query.gene.set}{A character vector of ensembl IDs}

\item{genomic.region.gene.sets}{A list where each element is a data.frame that includes at least SNPs and genes as ensembl IDs.}

\item{plot}{boolean, if true genrates a plot with the empirical null distribution and the observed overlapped given bye \code{query.gene.set}}
}
\value{
A GREAsummary data.frame
}
\description{
GREA.test
\code{GREA.test} plost the ditribution and observed overlap
}
