% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ib.traits.genes.list}
\alias{ib.traits.genes.list}
\title{List of 7 elements where each is a data.frame with gene sets defined by proximity to GWAS hits from Immunobase studies (results from immuno-chip platform)}
\format{A list with 7}
\usage{
data(ib.gw.traits.genes.list)
}
\description{
The list contains gene sets defined by proximity to GWAS hits for these diseases:
\itemize{
\item "CeD" Celiac Disease
\item "AS" Ankylosing Spondylitis
\item "MS" Multiple Sclerosis
\item "PBC" Primaire Biliaire Cholangitis
\item "RA" Rheumatoid Arthritis
\item "T1D" Rheumatoid Arthritis
\item "PSO" Psoriasis
}
}
\details{
Each data.frame of the list has the following information
\itemize{
\item chr
\item left
\item right
\item width
\item SNP
\item SNP.pos
}
}
\keyword{datasets,}
\keyword{gene}
\keyword{sets}
