% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{load.annotate.genes.from.GWAs.hits}
\alias{load.annotate.genes.from.GWAs.hits}
\title{Load and annotate GWAS summary stats
\code{load.annotate.genes.from.GWAs.hits} Given common summary stats}
\usage{
load.annotate.genes.from.GWAs.hits(file, source = "Immunobase",
  gene.window = 5e+05, gene.info)
}
\arguments{
\item{file}{Character, path to the file were the GWAS summary stats are located}

\item{source}{Character, either "GWASCatalog" or "Immunobase"}

\item{gene.window}{Numeric, genomic window}

\item{gene.info}{A data.frame with genomic coordinates for genes. For reference please look at data(gene.info)}
}
\value{
return ---
}
\description{
Load and annotate GWAS summary stats
\code{load.annotate.genes.from.GWAs.hits} Given common summary stats
}
