% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{null.dist.plotter}
\alias{null.dist.plotter}
\title{null.dist.plotter
\code{null.dist.plotter} plots the ditribution and observed overlap}
\usage{
null.dist.plotter(null.perc.overlap.per.gr, GREA.summary)
}
\arguments{
\item{null.perc.overlap.per.gr}{a list with the overlap between sampled genes and}

\item{GREA.summary}{a GREA.summary data.frame calculated through the \code{GREA.test} function}
}
\value{
a ggplot object
}
\description{
null.dist.plotter
\code{null.dist.plotter} plots the ditribution and observed overlap
}
