% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_filetype.R
\name{obtain_filetype}
\alias{obtain_filetype}
\title{Obtain Filetype}
\usage{
obtain_filetype(filelocation)
}
\arguments{
\item{filelocation}{A single string with the location and name of the file, e.g. "data/bla.csv"}
}
\value{
The filetype, e.g. "csv"
}
\description{
Obtains the filetype of a file in a string
}

