% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREA.functions.R
\name{pvalRight}
\alias{pvalRight}
\title{pvalRight
\code{pvalRight} given an empirical distribtion it calculates the quantile in which an observed measurement is located within the distribution}
\usage{
pvalRight(x, dist)
}
\arguments{
\item{x}{numeric, observed value to be compared against a null distribution}

\item{dist}{numeric vector defining a null distribution}
}
\value{
return the estimated pValue based on the null distrubtion provided
}
\description{
pvalRight
\code{pvalRight} given an empirical distribtion it calculates the quantile in which an observed measurement is located within the distribution
}
