% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_check.R
\name{wd_check}
\alias{wd_check}
\title{File-in-Working Directory checker}
\usage{
wd_check(filelocation)
}
\value{
A file location. Gives the file location without the WD if the file is in the WD, but returns only the input file location if it doesn't.
}
\description{
Function to determine if a file lies in a working directory.
}
\examples{
wd_check("/Users/stani/GitHub/GREA/data/csv_food_supply.csv")
}

