% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{GREA_read}
\alias{GREA_read}
\title{Read Data Wrapper}
\usage{
GREA_read(filelocation, sheetIndex = 1L, sep, dec = "auto", header)
}
\arguments{
\item{filelocation}{A single character string with the location and name of the file, e.g. \dQuote{data/bla.csv}.}

\item{sheetIndex}{An integer specifying which sheet to import from an Excel or ODS spreadsheet file.}

\item{sep}{A character string specifying a column separator for delimited files.}

\item{dec}{A character string specifying a decimal separator for delimited files.}

\item{header}{A logical indicating whether a delimited file contains a header row.}
}
\value{
A dataframe, containing the read-in data.
}
\description{
Reads data in any of the formats supported by \code{\link[rio]{import}}.
}

